renamed GameInfo2 to AssemblyViewer
