+++ UI Assets/AssemblyViewer/PropertyGetterPrivate.png
