+++ description.txt
@@ -10,8 +10,8 @@ Thanks to Pyroknight for drawing the sprites.
 Thanks to direwolf420, darthmorf and jopojelly for some UIElements.
 
 Latest Changes:
-v4.4 - GameInfo2
-    + Added GameInfo2
+v4.4 - AssemblyViewer
+    + Added AssemblyViewer
     + Fixed DownedBossToggle's UI
 
 v4.3 - added 1.4 features
