+++ src/Tools/AssemblyViewer/Elements/ClassButton.cs
@@ -4,7 +4,7 @@ using Terraria.GameContent.UI.Elements;
 using Terraria.ModLoader;
 using Terraria.UI;
 
-namespace Creativetools.src.Tools.GameInfo2.Elements
+namespace Creativetools.src.Tools.AssemblyViewer.Elements
 {
     class ClassButton : UIText
     {
@@ -13,7 +13,7 @@ namespace Creativetools.src.Tools.GameInfo2.Elements
         {
             _class = clas;
 
-            var texture = ModContent.Request<Texture2D>("Creativetools/UI Assets/GameInfo2/class", ReLogic.Content.AssetRequestMode.ImmediateLoad);
+            var texture = ModContent.Request<Texture2D>("Creativetools/UI Assets/AssemblyViewer/class", ReLogic.Content.AssetRequestMode.ImmediateLoad);
             Append(new UIImage(texture));
         }
 
@@ -21,7 +21,7 @@ namespace Creativetools.src.Tools.GameInfo2.Elements
         {
             base.Click(evt);
 
-            UISystem.UserInterface.SetState(new GameInfo2(_class.Namespace, _class));
+            UISystem.UserInterface.SetState(new AssemblyViewer(_class.Namespace, _class));
         }
     }
 }
